Parsing of config is done, validation of paths, port, host, client size and brackets

+++ ServerConfig.hpp
@@ -32,7 +32,7 @@ class ServerConfig
         ServerConfig();
         ~ServerConfig();
 
-        size_t pos;
+        int pos;
 
         // setters
         void    setPort(int port);
